# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Naveen Blog API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Post {
  id: string;
  title: string;
  slug: string;
  /** @nullable */
  excerpt?: string | null;
  content: string;
  category: string;
  /** @nullable */
  tags?: string[] | null;
  published: boolean;
  /** @nullable */
  cover_emoji?: string | null;
  /** @nullable */
  word_count?: number | null;
  /** @nullable */
  read_time?: number | null;
  created_at: string;
  updated_at: string;
}

export interface PostInput {
  title: string;
  content: string;
  category: string;
  custom_category?: string;
  excerpt?: string;
  tags?: string[];
  published?: boolean;
  cover_emoji?: string;
}

export interface PostUpdate {
  title?: string;
  content?: string;
  category?: string;
  custom_category?: string;
  excerpt?: string;
  tags?: string[];
  published?: boolean;
  cover_emoji?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type AuthResultUser = {
  id: string;
  email: string;
};

export interface AuthResult {
  success: boolean;
  user?: AuthResultUser;
}

export type AuthStatusUser = {
  id: string;
  email: string;
};

export interface AuthStatus {
  authenticated: boolean;
  user?: AuthStatusUser;
}

export type GetPostsParams = {
category?: string;
search?: string;
limit?: number;
};

export type GetPosts200 = {
  posts: Post[];
};

export type CreatePost201 = {
  post: Post;
};

export type GetPost200 = {
  post: Post;
};

export type UpdatePost200 = {
  post: Post;
};

export type DeletePost200 = {
  success: boolean;
};

export type AuthLogout200 = {
  success: boolean;
};