# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Naveen Blog API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: posts
    description: Blog post operations
  - name: auth
    description: Authentication operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /posts:
    get:
      operationId: getPosts
      tags: [posts]
      summary: List published posts
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: object
                properties:
                  posts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Post"
                required: [posts]
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new post (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                type: object
                properties:
                  post:
                    $ref: "#/components/schemas/Post"
                required: [post]
        "401":
          description: Unauthorized

  /posts/{slug}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a single post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                type: object
                properties:
                  post:
                    $ref: "#/components/schemas/Post"
                required: [post]
        "404":
          description: Not found
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a post (admin only)
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUpdate"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                type: object
                properties:
                  post:
                    $ref: "#/components/schemas/Post"
                required: [post]
        "401":
          description: Unauthorized
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post (admin only)
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required: [success]
        "401":
          description: Unauthorized

  /auth/login:
    post:
      operationId: authLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials

  /auth/logout:
    post:
      operationId: authLogout
      tags: [auth]
      summary: Admin logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required: [success]

  /auth/me:
    get:
      operationId: authMe
      tags: [auth]
      summary: Check auth status
      responses:
        "200":
          description: Auth status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatus"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Post:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: ["string", "null"]
        content:
          type: string
        category:
          type: string
        tags:
          type: ["array", "null"]
          items:
            type: string
        published:
          type: boolean
        cover_emoji:
          type: ["string", "null"]
        word_count:
          type: ["integer", "null"]
        read_time:
          type: ["integer", "null"]
        created_at:
          type: string
        updated_at:
          type: string
      required: [id, title, slug, content, category, published, created_at, updated_at]

    PostInput:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
        custom_category:
          type: string
        excerpt:
          type: string
        tags:
          type: array
          items:
            type: string
        published:
          type: boolean
        cover_emoji:
          type: string
      required: [title, content, category]

    PostUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
        custom_category:
          type: string
        excerpt:
          type: string
        tags:
          type: array
          items:
            type: string
        published:
          type: boolean
        cover_emoji:
          type: string

    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]

    AuthResult:
      type: object
      properties:
        success:
          type: boolean
        user:
          type: object
          properties:
            id:
              type: string
            email:
              type: string
          required: [id, email]
      required: [success]

    AuthStatus:
      type: object
      properties:
        authenticated:
          type: boolean
        user:
          type: object
          properties:
            id:
              type: string
            email:
              type: string
          required: [id, email]
      required: [authenticated]