# lib/api-zod/src/generated/types/post.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Naveen Blog API
 * OpenAPI spec version: 0.1.0
 */

export interface Post {
  id: string;
  title: string;
  slug: string;
  /** @nullable */
  excerpt?: string | null;
  content: string;
  category: string;
  /** @nullable */
  tags?: string[] | null;
  published: boolean;
  /** @nullable */
  cover_emoji?: string | null;
  /** @nullable */
  word_count?: number | null;
  /** @nullable */
  read_time?: number | null;
  created_at: string;
  updated_at: string;
}