# lib/api-zod/src/generated/types/postInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Naveen Blog API
 * OpenAPI spec version: 0.1.0
 */

export interface PostInput {
  title: string;
  content: string;
  category: string;
  custom_category?: string;
  excerpt?: string;
  tags?: string[];
  published?: boolean;
  cover_emoji?: string;
}